Better tests on float01.
	* mln/value/float01.hh: Fix.
	* mln/value/float01_.hh: Fix.
	* tests/value_float01.cc: Update.

git-svn-id: https://svn.lrde.epita.fr/svn/oln/trunk@1221 4aad255d-cdde-0310-9447-f3009e2ae8c0

+++ milena/mln/value/float01.hh
@@ -146,7 +146,7 @@ namespace mln
     double float01::value() const
     {
       mln_invariant(nbits_ != 0);
-      return val_ / internal::two_pow_n_minus_1(nbits_);
+      return double(val_) / internal::two_pow_n_minus_1(nbits_);
     }
 
     unsigned long float01::value_ind() const
